Included image commit versions and metedata badges
Updated image URLs for jsdelivr CDN distribution to include commit
versions and included shield.io metadata badges.

Jsdelivr was not able to successfully locate the screenshot with the
direct Github URI - this was addressed by including the commit version.
The icon URI was also updated to include the commit version for best
practice/consistency.