Nuspec modifications for Chocolatey integration
Updated elements of the nuspec to better integrate with the Chocolatey
repository.  The package information was not rendering properly on the
Chocolatey site, it was inappropriate to include the metadata badges and
the package was failing automated checks as it was failing to install due
to a missing dependency.

Updates were made to:
 - included version for icon CDN delivery
 - remove the image link which was not being rendered correctly
 - remove the metadata badges
 - collapse the initial parts of the description onto a single line as
   it was not being rendered correctly
 - added an additional dependency for vscode