Updated version and documentation handling
Version handling needed to be updated to remove the explicit replacement
of the version in the nuspec as this is handled already by the au
updater and to include an update for the version in the dependency
on VS Code.

Documentation handing was modified to address an apparent issue in the
AU updater.  Prior nuspec changes were being dropped when the
description was being updated through the Set-DescriptionFromReadme
function - search and replace results were being reverted when the
nuspec was written following on frrom the description update from the
readme.

As an interim measure the use of the automated update has been disabled
but the correct action is to address the bug upstream in the AU updater.