Updated license references - fixes gh-1
Updated license references to use those in the extension distribution
rather than the source code repository.  This was to address the
possibility that the licensing for the binary distribution and the
source code may diverge over time.  Given it is the binary which is
being redistributed the license needed to be aligned.