(gh-30) Removed explicit VSCode dependencies
VSCode extension packages had an explicit dependency on a version
of Visual Studio Code.  This removed choice from users as it forces the
install of a VSCode version which may not be inline with user intent, especially if the user is employing an alternate distribution such as vscode-
insiders or vscodium.

Updated packages to remove the dependency and shift the version
identification requirements to documentation.