(gh-83) Updated URL matched for version in update script
The URL used to match against when determining the version is empty so
the version matching will return nothing.  To ensure a version is
returned when the update script is run the version match needs to be
performed against a populated URL.

Updated the URL used for the match to $url64 to achieve this.