(gh-85) Remove copyright to address package validation failures
The package validation for haku-neko is failing due the minimum
copyright character count check.  As the block was only popuylated
with 4 characters the rule was being triggered.

With the application being covered under an UNLICENSE which explicitly
revokes any claim to copyright there is in fact none in place.  Given this
the appropriate fix was to remove the copyright blocks from the package
specifications.