(gh-93) Corrected icon URL in package definition
The package definition was not correct due to the non-locateable icon.
The icon was not locatable due to an error in the URL.  To correct the
package definition the URL was corrected so that the icon could be
accessed at the URL.

The documentation was also updated to remove a linebreak such that the
display was clean.