(gh-98) Added meta-package for tldr
Added a meta-package for tldr.  The purpose of providing a meta-package
was to abstract the base tldr package from knowledge of a specific
client.  Given the multiple clients available capabilities and
preferences may change over time so the flexibility in being able to
select a particular tldr client was desired.

For the initial implementation the Windows  tldr++ client was selected.