(gh-102) Removed version specification from dependency
The definition of the tldr++ client dependency was incorrect.  With an
exclusive minimum version matching the desired dependency version it
was not being located successfully.

Given the behaviour of the Chocolatey internal Nuspec version for an
unspecified dependency version is to locate the highest available
version there is in fact no need for an explicit version so this was
removed.