(gh-105) Corrected checksum updates
The checksums for the embedded installer were not updating correctly on
package update.  Without a checksum update the package install was
failing due to the bad checksum in place.

Checksums were not being updated as the regex block to match on
Checksum64 for the chocolateyinstall.ps1 was not present in the update
script.  The script needed to be modifed and the checksum patched to
address this.