(gh-134) Corrected update script for Keys package
Package updates were  failing due to incorrect file tokens being
populated.  The AU scripts are not able to locate the filename as a
result and a blank valuewas being used.

Given the package supports both 32 and 64-bit versions in a single msi
updated the file tokens so that they would be picked up by AU (use of
FileName32) and corrected the validation and install files with the
appropriate values.