(gh-138) Corrected GitHub Pull Requests extension pack updater
The package updater for the GitHub Pull Requests and Issues VSCode
extesnion was pulling in a bad version which did not exist.  This was
due to an error in the name of the extension used in the update script
with a version being pulled in for a completely unrelated extension.

Corrected the extension used in the updater and ensured that version
references were aligned accordingly.