(gh-155) Improved copyright handling
Updated handling to deal with the first published and last updated being
equivalent and produce a single date (eg. 2020) rather than a repeated
year range (eg. 2020 - 2020).

Provided support for mapping copyright strings to arrive at something
better suited for display.  The exising precendence for copyright
information is still followed (author --> publisher display name -->
publisher) with an ability to set overrides at each of these leves in
configuration (eg. golang --> Google LLC).  At this stage the
configuration is embedded in the module and it might make sense to break
this out in future.