(gh-164) Corrected Keys automatic updater
The automatic updater for the Keys package was failing due to case
insensitivity on Powershell regex additional characters are being matched
in the URL when updating the VERIFICATION.txt file. This wass causing
the direct download URL to be truncated and resulting in  failures on
subsequent updates.

In addition to correcting the update script the package was updated to
the next to latest version so that the previous version could be pushed
manually and the automatic update verified with the next package push.