(gh-209) Fix CLOC Automatic Update
The AU automatic update for the CLOC package was failing due to a
mismatch in the regular expression extracting the url.  An empty URL was
being extracted which ensured that the version number was not able to be
extracted.  The lack of the version number was breaking the update.

Corrected the URL extraction to provide a valid URL and ensure the
availability of the version number.