(gh-213) Correct Prime95 Bugtracker Url
The Prime95 package was failing validation due to an invalid URL in the
Bugtracker element. The element is currently an email address which is
not supported by the package validator.  Removed the Bugtracker element
as there is no alternative to email available - Prime95 does not have a
public bug tracker.

There was also an issue with the automatic update picking up unintended
additional version strings.  Reverted these version strings to the
correctvlegacy version and modified the update script to ensure that
they would not be picke up in future.