(gh-152) SQL Server Compact Package Updates
Minor updates to the SQL Server Compact package.  Modified the grammar
used for the lanaguage example, removed the CI build status link given
the package is manual,  updated formatting in the verification instructions
and amended the install script to use a struct versus class for compatability
with earlier Poweshell versions.