(gh-248) Ensure correct version resolution on update
A number of packages were reporting errors on checking for updates - the
correct versions were not being resolved on update checks.  There were
2 separate categories here - handling of updates for packages where no
windows version was provided and errors in the matching of the URLs on
the download page for version determination.

The pacakges sd-cli, tldr-plusplus amd watchman were in category 1 while
beekeeper-studio, kui and mondodb-shell were in category 2.

Updated handling or matching to deal with each of these scenarios
cleanly.