(gh-254) Flashpoint Infinity update error
The automated updater is failing on extracting the archive name from the
URL. The failure is occurring due to an order of precedence in the
expression with a parameter binding before evaluation of the LHS string
evaluation.

Update the string evaluation with parentheses to ensure the expected
order of evaluation.