(gh-263) Cater to additional patch versions
Update to the regular expression used for version matching to cater to
patch versions eg. a.b.c.d when making updates.  An additional patch
version was introduced to correct an issue with the package and this
change will ensure that automated updates are able to locate the patch
version and revert to the standard versioning scheme of a.b.c.