(gh-200) Support user installs
Chocolatey packages are normally installed at the machine level with
administrative privileges.  This may not always be the case and the ant
package would fail where this was not the case
due to the use of an environment variable at the
machine level.

Added a 'User' parameter to support package installs where admin is not
available.