(gh-283) Use Windows profile directory for HOME
The rsync application used to require a HOME environment variable to be set.
This was changed with the 6.2.1 release which now utilises the Windows profile
directory as HOME.

With this being the case there is no longer a requirement to set the
CWRSYNC_HOME environment variable or use the batch wrapper - the rsync.exe
file may now be called directly.