(gh-305) Include uninstall script in automatic updates
The chocolateyUninstall.ps1 file wasnot  included in the automatic
updates.  This resulted in a stale version being left in the uninstall
script which referred to a prior zip file on the uninstall.

The automatic update script was modified to include the uninstall
script and the version synchronized with the current package version.