(gh-304) Update version extraction
Updates were failing as the regex was not matching a version in the URL
when the au_GetLatest script was called. called.  With no version number
being extracted the package update is failing.

Updated version regex to correctly match versions and included handling
to append a '.0' to the extracted version if only an integral value was
returned.