(gh-371) Correct package update script
The package update script was not correctly handling updates due to the
portable package filename being mapped from the intended .zip file to a
.msi file.

This was occuring on file retrieval in AU due to a fallback
to a .msi file.  With no FileType being specified the downloaded file was
being saved as a .msi rather than a .zip and this was propagated
forward which resulted in an incorrect filename being used.

An explicit filetype needed to be specified in the case of the zip file.