(gh-370) Correct names used for shim generation
Shims for the package were being installed with an additional .exe
appended.  The names used for shim generation needed updating to remove
the .exe as this is added automatically in Install-BinFile - if a file
already has the .exe extension it would be added again resulting in a
shim ending with .exe.exe.