(gh-388) Update filename token replacement
Package installation was failing due to an invalid filename in the
install script.  The name was incorrect as the replacment token used for
the filename when updating the install script was incorrect - it was
leaving the existing filename in place while appending the new.

Updated the token replacement to only use the new filename when updating
the install script.