(gh-393) Reverted documentation URL to v0.35.1
Package validation was failing due to the DocsUrl element no longer
being resolvable.  This was due to an interim build being published for
which no tag was created.

Currently no clarity on the intended path forward - tags may have been
abandonded or it could just be an error for this build so pinning the
documentation to that for the previous (v0.35.1) published version for
now.