(gh-410) Updated to reflect build changes since 0.9.4
The automatic updater for the Gephi package was failing since the 0.9.5
update due to a change of teh distribtion files - x32 and x64 installers
are now produced as build artifacts and the automatic update was failing
as a result.

Modified regular expressions to ensure only a single matchand updated
documentation to reflect changes to the package which no longer requires
an external JDK to be available on install.