(gh-411) Update package documentation reference
Package validation was failing due to an invalid documentation URL - the
versioning scheme for the distribtion had been updated so that it was no
longer synchronized with the tagging mechanism on the Github repository.

Updated to use a standard documetnation URL that was persistent across
versions and no longer associated with the Github tagged versions and
refreshed the regular expression usage.