(gh-363) Correct package upgrade failures
Package upgrades were failing due to the presence of the install
directory/contents which cause the new install to fail.  The directory
is present as the Chocoatey auto-uninstaller is not correctly
uninstalling the package - the Qt installer does not appear to be
handled correctly by Chocolatey.

To address this automatic uninstall is being skipped and a specific
uninstall script which correctly handles the uninstall introduced.