(gh-431) Modified package updates to handle GitHub structure
All GitHub packages were broken due to a structural change on the
presentation of assets associated with a tag.  With the change to the
tag the AU update scripts were no longer capable of determining the
version.

Modified the handling to successfuly retrieve asset versions from GitHub
to repair automated version checks/package updates.