(gh-441) Correct sliksvn package updates
The sliksvn package was no longer updating due to a change in the
resource format - the package binary distribution had changed from raw
msi files to msi files additionally packaged in zip files.

Modified the updater to successfully retrieve and handle the msi files
packaged in zip files and enhanced the package to utilise shims to
provide a defaut option of not adding the install\bin directory to the
path.