(gh-466) Add helper functions to assist with binary version detection
Some binaries are made available from a consistent URL regardless of the
version.

To assist with processing these binaries a mechanism to check for an
update is required to limit the need to download binaries to check for
updates. This can be achieved using the ETag header.

Added helper functions to assist handling the ETag header by persisting
values/detecting changes.