(gh-473) Amended new package detection to handle x86 and x64 architectures
Package build was failing as the URL syntax was invalid - no URL was
being retrieved from the JSON describing the dotCover-cli binaries.

The package and JSON were updated to include an additional package type
on windows - what was  previously known as windows has been split
and now there are windows32 and windows64 versions representing the
x86 and x64 architectures respectively. Previously only the x64
architecture was provided and represented as windows.

Modified package update to handle this,  consolidated the regular
expressions and added support for the new 32-bit binaries.